requirements.txt
# ArquimedesAI v1.2 - Modern Tech Stack Requirements
# Based on latest stable versions (October 2025)

# Core Framework
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Vector Store & Embeddings
langchain-qdrant>=0.2.0  # Official Qdrant integration (replaces deprecated langchain_community.vectorstores.Qdrant)
qdrant-client>=1.12.0
langchain-huggingface>=0.1.0

# Retrieval & Reranking
rank-bm25>=0.2.2  # BM25 sparse retrieval for hybrid search
sentence-transformers>=2.2.0  # Cross-encoder reranking

# Document Processing
docling>=2.0.0
langchain-docling>=0.1.0  # Official Docling integration with HybridChunker

# LLM Integration
langchain-ollama>=0.3.10  # Ollama integration (replaces deprecated langchain_community.llms.Ollama)

# Configuration
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# Discord Bot
discord.py>=2.4.0

# CLI
typer>=0.15.0
rich>=13.0.0

# Text Processing
langchain-text-splitters>=0.3.0

# Optional: Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Optional: Code Quality
ruff>=0.8.0
mypy>=1.13.0

